encoder/requirements_agnomin.txt
matplotlib==3.3.2
pandas==1.1.3
pycparser==2.20
networkx==2.5
tqdm==4.50.2
scikit-learn==0.23.2
numpy==1.16.4
six==1.12
requests==2.24.0
h5py==2.10.0
numba==0.51.2
pygraphviz
kmeans-pytorch
transformers
torchmetrics